chore: integrate zero length string test into other tests
