+++ path/normalize_test.ts
@@ -1,0 +1,10 @@
+// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
+import { assertEquals } from "../assert/mod.ts";
+import { normalize } from "./normalize.ts";
+
+Deno.test(`normalize() returns "." if input is empty`, function () {
+  assertEquals(normalize(""), ".");
+
+  const pwd = Deno.cwd();
+  assertEquals(normalize(pwd), pwd);
+});
